+++ NEWS.md
@@ -6,5 +6,6 @@ First public release.
  * Varying change points
  * Basic GLM: Gaussian, binomial, Bernoulli, and Poisson, and associated vignettes.
  * summary(fit), fixef(fit), and ranef(fit)
- * plot(fit, "segments") and plot(fit, "bayesplot-name-here")
- * Basic unit tests to ensure non-breaking code for a wide variety of models.
+ * plot(fit, "segments") and plot(fit, "bayesplot-name-here") with some options
+ * 1000+ basic unit tests to ensure non-breaking code for a wide variety of models.
+ * Testing and model comparison using `loo` and `hypothesis`